organized backend and frontend folders
