+++ friendster-frontend/vite-project/.eslintrc.cjs
