+++ friendster-frontend/vite-project/.gitignore
