+++ friendster-frontend/vite-project/README.md
