+++ friendster-frontend/vite-project/index.html
