+++ friendster-frontend/vite-project/package.json
