+++ friendster-frontend/vite-project/public/friendster.png
