+++ friendster-frontend/vite-project/src/App.css
