+++ friendster-frontend/vite-project/src/App.jsx
