+++ friendster-frontend/vite-project/src/assets/Friendster.png
