+++ friendster-frontend/vite-project/src/assets/stylesheets/_navbar.scss
