+++ friendster-frontend/vite-project/src/assets/stylesheets/index.scss
