+++ friendster-frontend/vite-project/src/main.jsx
