+++ friendster-frontend/vite-project/vite.config.js
