+++ frienster-backend/app.js
@@ -1,4 +1,4 @@
-//add A EXPRESS APP JS 
+//add A EXPRESS APP JS
 const express = require("express");
 const bodyParser = require("body-parser");
 const admin = require("firebase-admin");
