+++ frienster-backend/middleware/auth.js
