+++ frienster-backend/package.json
@@ -35,5 +35,8 @@
     "firebase-admin": "^11.11.0",
     "nodemon": "^3.0.1",
     "pg": "^8.11.3"
+  },
+  "devDependencies": {
+    "create-vite": "^4.4.1"
   }
 }
