+++ frienster-backend/routes/registrationRoutes.js
